+++ PSFramework/internal/configurations/serialization.ps1
@@ -1,2 +1,2 @@
 ﻿# The path where type-files are stored when registered
-Set-PSFConfig -Module PSFramework -Name 'Serialization.WorkingDirectory' -Value "$env:APPDATA\WindowsPowerShell\PSFramework\TypeData" -Initialize -Validation "string" -Description "The folder in which registered type extension files are placed before import. Relevant for Register-PSFTypeSerializationData."
+Set-PSFConfig -Module PSFramework -Name 'Serialization.WorkingDirectory' -Value $script:path_typedata -Initialize -Validation "string" -Description "The folder in which registered type extension files are placed before import. Relevant for Register-PSFTypeSerializationData."