+++ PSFramework/internal/scripts/postimport.ps1
@@ -1,49 +1,52 @@
-﻿# Initialize the configurations
-foreach ($file in (Get-ChildItem -Path "$PSModuleRoot\internal\configurationvalidation\*.ps1"))
+﻿Import-PSFCmdlet -Name Write-PSFMessage -Type ([PSFramework.Commands.WritePSFMessageCommand]) -HelpFile "$($PSModuleRoot)$($dc)en-us$($dc)PSFramework.dll-Help.xml" -Module $ExecutionContext.SessionState.Module
+Import-PSFCmdlet -Name Set-PSFConfig -Type ([PSFramework.Commands.SetPSFConfigCommand]) -HelpFile "$($PSModuleRoot)$($dc)en-us$($dc)PSFramework.dll-Help.xml" -Module $ExecutionContext.SessionState.Module
+
+# Initialize the configurations
+foreach ($file in (Get-ChildItem -Path "$($PSModuleRoot)$($dc)internal$($dc)configurationvalidation$($dc)*.ps1"))
 {
 	. Import-ModuleFile -Path $file.FullName
 }
-foreach ($file in (Get-ChildItem -Path "$PSModuleRoot\internal\configurations\*.ps1"))
+foreach ($file in (Get-ChildItem -Path "$($PSModuleRoot)$($dc)internal$($dc)configurations$($dc)*.ps1"))
 {
 	. Import-ModuleFile -Path $file.FullName
 }
 
 # Import configuration settings from registry
-. Import-ModuleFile -Path "$PSModuleRoot\internal\scripts\loadConfigurationFromRegistry.ps1"
+. Import-ModuleFile -Path "$($PSModuleRoot)$($dc)internal$($dc)scripts$($dc)loadConfigurationPersisted.ps1"
 
 # Load each logging provider
-foreach ($file in (Get-ChildItem -Path "$PSModuleRoot\internal\loggingProviders\"))
+foreach ($file in (Get-ChildItem -Path "$($PSModuleRoot)$($dc)internal$($dc)loggingProviders$($dc)"))
 {
 	. Import-ModuleFile -Path $file.FullName
 }
 
 # Start the logging system
-. Import-ModuleFile -Path "$PSModuleRoot\internal\scripts\async-logging2.ps1"
+. Import-ModuleFile -Path "$($PSModuleRoot)$($dc)internal$($dc)scripts$($dc)async-logging2.ps1"
 
 # Launch the Tab Expansion system
-foreach ($file in (Get-ChildItem -Path "$PSModuleRoot\internal\tepp\scripts\"))
+foreach ($file in (Get-ChildItem -Path "$($PSModuleRoot)$($dc)internal$($dc)tepp$($dc)scripts$($dc)"))
 {
 	. Import-ModuleFile -Path $file.FullName
 }
-. Import-ModuleFile -Path "$PSModuleRoot\internal\tepp\tepp-assignment.ps1"
+. Import-ModuleFile -Path "$($PSModuleRoot)$($dc)internal$($dc)tepp$($dc)tepp-assignment.ps1"
 
 # Load parameter class extensions
-foreach ($file in (Get-ChildItem -Path "$PSModuleRoot\internal\parameters\" -Filter "*.ps1"))
+foreach ($file in (Get-ChildItem -Path "$($PSModuleRoot)$($dc)internal$($dc)parameters$($dc)" -Filter "*.ps1"))
 {
 	. Import-ModuleFile -Path $file.FullName
 }
 
 # Import the aliases for PSFramework types
-. Import-ModuleFile -Path "$PSModuleRoot\bin\type-aliases.ps1"
+. Import-ModuleFile -Path "$($PSModuleRoot)$($dc)bin$($dc)type-aliases.ps1"
 
 # Register the task engine
-. Import-ModuleFile -Path "$PSModuleRoot\internal\scripts\taskEngine.ps1"
+. Import-ModuleFile -Path "$($PSModuleRoot)$($dc)internal$($dc)scripts$($dc)taskEngine.ps1"
 
 # Register the unimport reaction
-. Import-ModuleFile -Path "$PSModuleRoot\internal\scripts\removalEvent.ps1"
+. Import-ModuleFile -Path "$($PSModuleRoot)$($dc)internal$($dc)scripts$($dc)removalEvent.ps1"
 
 # Finally register the license
-. Import-ModuleFile -Path "$PSModuleRoot\internal\scripts\license.ps1"
+. Import-ModuleFile -Path "$($PSModuleRoot)$($dc)internal$($dc)scripts$($dc)license.ps1"
 
 #region Declare runtime variable for the flow control component
 $paramNewVariable = @{