+++ library/PSFramework/Commands/WritePSFMessageCommand.cs
@@ -116,6 +116,12 @@ namespace PSFramework.Commands
         /// </summary>
         [Parameter()]
         public bool EnableException;
+
+        /// <summary>
+        /// Enables breakpoints on the current message. By default, setting '-Debug' will NOT cause an interrupt on the current position.
+        /// </summary>
+        [Parameter()]
+        public SwitchParameter Breakpoint;
         #endregion Parameters
 
         #region Private fields
@@ -183,6 +189,11 @@ else { Write-PSFHostColor -String $string -DefaultColor ([PSFramework.Message.Me
         /// List of tags to process
         /// </summary>
         private List<string> _Tags = new List<string>();
+
+        /// <summary>
+        /// Whether debug mode is enabled
+        /// </summary>
+        private bool _isDebug;
         #endregion Private fields
 
         #region Private properties
@@ -247,6 +258,34 @@ else { Write-PSFHostColor -String $string -DefaultColor ([PSFramework.Message.Me
                     return GetMessageColor();
             }
         }
+
+        /// <summary>
+        /// Provide breadcrumb queue of the callstack
+        /// </summary>
+        private string _BreadCrumbsString
+        {
+            get
+            {
+                string crumbs = String.Join(" > ", _callStack.Select(name => name.FunctionName).Where(name => name != "Write-PSFMessage" && name != "Stop-PSFFunction" && name != "<ScriptBlock>").Reverse().ToList());
+                if (crumbs.EndsWith(FunctionName))
+                    return String.Format("[{0}]\n    ", crumbs);
+                return String.Format("[{0}] [{1}]\n    ", crumbs, FunctionName);
+            }
+        }
+
+        /// <summary>
+        /// Provide a breadcrumb queue of the callstack in color tags
+        /// </summary>
+        private string _BreadCrumbsStringColored
+        {
+            get
+            {
+                string crumbs = String.Join("</c> > <c='sub'>", _callStack.Select(name => name.FunctionName).Where(name => name != "Write-PSFMessage" && name != "Stop-PSFFunction" && name != "<ScriptBlock>").Reverse().ToList());
+                if (crumbs.EndsWith(FunctionName))
+                    return String.Format("[<c='sub'>{0}</c>]\n    ", crumbs);
+                return String.Format("[<c='sub'>{0}</c>] [<c='sub'>{1}</c>]\n    ", crumbs, FunctionName);
+            }
+        }
         #endregion Private properties
 
         #region Cmdlet Implementation
@@ -303,6 +342,8 @@ else { Write-PSFHostColor -String $string -DefaultColor ([PSFramework.Message.Me
             if (Tag != null)
                 foreach (string item in Tag)
                     _Tags.Add(item);
+
+            _isDebug = (_callStack.Count() > 1) && _callStack.ElementAt(_callStack.Count() - 2).InvocationInfo.BoundParameters.ContainsKey("Debug");
             #endregion Resolving Meta Information
         }
 
@@ -437,20 +478,40 @@ else { Write-PSFHostColor -String $string -DefaultColor ([PSFramework.Message.Me
             {
                 if ((_callStack.Count() > 1) && _callStack.ElementAt(_callStack.Count() - 2).InvocationInfo.BoundParameters.ContainsKey("Verbose"))
                     InvokeCommand.InvokeScript(@"$VerbosePreference = 'Continue'");
+                    //SessionState.PSVariable.Set("VerbosePreference", ActionPreference.Continue);
 
-                WriteVerbose(_MessageStreams);
+                    WriteVerbose(_MessageStreams);
                 channels = channels | LogEntryType.Verbose;
             }
 
             if ((MessageHost.MaximumDebug >= (int)Level) && (MessageHost.MinimumDebug <= (int)Level))
             {
-                WriteDebug(_MessageStreams);
-                channels = channels | LogEntryType.Debug;
+                bool restoreInquire = false;
+                if (_isDebug)
+                {
+                    if (Breakpoint.ToBool())
+                        InvokeCommand.InvokeScript(false, ScriptBlock.Create(@"$DebugPreference = 'Inquire'"), null, null);
+                    else
+                    {
+                        InvokeCommand.InvokeScript(false, ScriptBlock.Create(@"$DebugPreference = 'Continue'"), null, null);
+                        restoreInquire = true;
+                    }
+                    WriteDebug(String.Format("{0} | {1}", Line, _MessageStreams));
+                    channels = channels | LogEntryType.Debug;
+                }
+                else
+                {
+                    WriteDebug(_MessageStreams);
+                    channels = channels | LogEntryType.Debug;
+                }
+
+                if (restoreInquire)
+                    InvokeCommand.InvokeScript(false, ScriptBlock.Create(@"$DebugPreference = 'Inquire'"), null, null);
             }
             #endregion Message handling
 
             #region Logging
-            LogEntry entry = LogHost.WriteLogEntry(_MessageSystem, channels, _timestamp, FunctionName, ModuleName, _Tags, Level, System.Management.Automation.Runspaces.Runspace.DefaultRunspace.InstanceId, Environment.MachineName, File, Line, Target);
+            LogEntry entry = LogHost.WriteLogEntry(_MessageSystem, channels, _timestamp, FunctionName, ModuleName, _Tags, Level, System.Management.Automation.Runspaces.Runspace.DefaultRunspace.InstanceId, Environment.MachineName, File, Line, _callStack, String.Format("{0}\\{1}",Environment.UserDomainName, Environment.UserName), Target);
             #endregion Logging
 
             foreach (MessageEventSubscription subscription in MessageHost.Events.Values)
@@ -573,7 +634,19 @@ else { Write-PSFHostColor -String $string -DefaultColor ([PSFramework.Message.Me
         {
             if (!String.IsNullOrEmpty(_message))
                 return _message;
-            _message = String.Format("[{0}][{1}] {2}", _timestamp.ToString("HH:mm:ss"), FunctionName, GetMessageSimple());
+            if (MessageHost.EnableMessageTimestamp && MessageHost.EnableMessageBreadcrumbs)
+                _message = String.Format("[{0}]{1}{2}", _timestamp.ToString("HH:mm:ss"), _BreadCrumbsString, GetMessageSimple());
+            else if (MessageHost.EnableMessageTimestamp && MessageHost.EnableMessageDisplayCommand)
+                _message = String.Format("[{0}][{1}] {2}", _timestamp.ToString("HH:mm:ss"), FunctionName, GetMessageSimple());
+            else if (MessageHost.EnableMessageTimestamp)
+                _message = String.Format("[{0}] {1}", _timestamp.ToString("HH:mm:ss"), GetMessageSimple());
+            else if (MessageHost.EnableMessageBreadcrumbs)
+                _message = String.Format("{0}{1}", _BreadCrumbsString, GetMessageSimple());
+            else if (MessageHost.EnableMessageDisplayCommand)
+                _message = String.Format("[{0}] {1}", FunctionName, GetMessageSimple());
+            else
+                _message = GetMessageSimple();
+
             return _message;
         }
 
@@ -603,7 +676,19 @@ else { Write-PSFHostColor -String $string -DefaultColor ([PSFramework.Message.Me
             if (!String.IsNullOrEmpty(_messageColor))
                 return _messageColor;
 
-            _messageColor = String.Format("[<c='sub'>{0}</c>][<c='sub'>{1}</c>] {2}", _timestamp.ToString("HH:mm:ss"), FunctionName, _errorQualifiedMessage);
+            if (MessageHost.EnableMessageTimestamp && MessageHost.EnableMessageBreadcrumbs)
+                _messageColor = String.Format("[<c='sub'>{0}</c>]{1} {2}", _timestamp.ToString("HH:mm:ss"), _BreadCrumbsStringColored, _errorQualifiedMessage);
+            else if (MessageHost.EnableMessageTimestamp && MessageHost.EnableMessageDisplayCommand)
+                _messageColor = String.Format("[<c='sub'>{0}</c>][<c='sub'>{1}</c>] {2}", _timestamp.ToString("HH:mm:ss"), FunctionName, _errorQualifiedMessage);
+            else if (MessageHost.EnableMessageTimestamp)
+                _messageColor = String.Format("[<c='sub'>{0}</c>] {1}", _timestamp.ToString("HH:mm:ss"), _errorQualifiedMessage);
+            else if (MessageHost.EnableMessageBreadcrumbs)
+                _messageColor = String.Format("{0}{1}", _BreadCrumbsStringColored, _errorQualifiedMessage);
+            else if (MessageHost.EnableMessageDisplayCommand)
+                _messageColor = String.Format("[<c='sub'>{0}</c>] {1}", FunctionName, _errorQualifiedMessage);
+            else
+                _messageColor = _errorQualifiedMessage;
+            
             return _messageColor;
         }